pyproject.toml
[project]
name = "castaway"
version = "0.1.0"
description = "MCP server for interacting with Ethereum applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "web3>=7.11.1",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.3.5",
    "sqlalchemy>=2.0.0",
]
